+++ src/ui/main/apply_patch.rs
@@ -5,7 +5,7 @@ use crate::i18n::*;
 
 use super::{App, AppMsg};
 
-pub fn apply_patch<T: PatchExt + Send + Sync + 'static>(sender: ComponentSender<App>, patch: T) {
+pub fn apply_patch(sender: ComponentSender<App>, patch: PlayerPatch, rename_mhypbase: bool) {
     match patch.status() {
         PatchStatus::NotAvailable |
         PatchStatus::Outdated { .. } |
@@ -18,8 +18,6 @@ pub fn apply_patch<T: PatchExt + Send + Sync + 'static>(sender: ComponentSender<
             let config = Config::get().unwrap();
 
             std::thread::spawn(move || {
-                let mut apply_patch_if_needed = true;
-
                 if let Err(err) = patch.apply(config.game.path.for_edition(config.launcher.edition), config.patch.root) {
                     tracing::error!("Failed to patch the game");
 
@@ -27,16 +25,29 @@ pub fn apply_patch<T: PatchExt + Send + Sync + 'static>(sender: ComponentSender<
                         title: tr("game-patching-error"),
                         description: Some(err.to_string())
                     });
+                }
 
-                    // Don't try to apply the patch after state updating
-                    // because we just failed to do it
-                    apply_patch_if_needed = false;
+                else if rename_mhypbase {
+                    let game_folder = config.game.path.for_edition(patch.edition);
+
+                    let mhypbase = game_folder.join("mhypbase.dll");
+                    let mhypbase_bak = game_folder.join("mhypbase.dll.bak");
+
+                    if mhypbase.exists() {
+                        if let Err(err) = std::fs::rename(mhypbase, mhypbase_bak) {
+                            tracing::error!("Failed to rename mhypbase file");
+
+                            sender.input(AppMsg::Toast {
+                                title: tr("game-patching-error"),
+                                description: Some(err.to_string())
+                            });
+                        }
+                    }
                 }
 
                 sender.input(AppMsg::DisableButtons(false));
                 sender.input(AppMsg::UpdateLauncherState {
                     perform_on_download_needed: false,
-                    apply_patch_if_needed,
                     show_status_page: true
                 });
             });