+++ src/ui/main/download_diff.rs
@@ -68,7 +68,6 @@ pub fn download_diff(sender: ComponentSender<App>, progress_bar_input: Sender<Pr
         sender.input(AppMsg::SetDownloading(false));
         sender.input(AppMsg::UpdateLauncherState {
             perform_on_download_needed,
-            apply_patch_if_needed: false,
             show_status_page: false
         });
     });