+++ src/ui/main/migrate_folder.rs
@@ -19,7 +19,6 @@ pub fn migrate_folder(sender: ComponentSender<App>, from: PathBuf, to: PathBuf, 
         sender.input(AppMsg::DisableButtons(false));
         sender.input(AppMsg::UpdateLauncherState {
             perform_on_download_needed: false,
-            apply_patch_if_needed: false,
             show_status_page: true
         });
     });